Implement ticket management functionality with CRUD operations and user authorization
