+++ 404tickets/backend/middlewares/logger.js
@@ -1,0 +1,14 @@
+const logger = (req, res, next) => {
+  const start = Date.now();
+  
+  res.on('finish', () => {
+    const duration = Date.now() - start;
+    console.log(
+      `${new Date().toISOString()} | ${req.method} ${req.originalUrl} | ${res.statusCode} | ${duration}ms`
+    );
+  });
+  
+  next();
+};
+
+module.exports = logger;
