+++ 404tickets/backend/models/User.js
@@ -49,4 +49,3 @@ UserSchema.methods.matchPassword = async function(enteredPassword) {
 };
 
 module.exports = mongoose.model('User', UserSchema);
-models
